Moved displayArrayToScreen and hangmanImages files to components directory
