+++ assets/js/components/hangmanImages.js
@@ -11,13 +11,3 @@ export const hangmanImages = () => {
 
   id.value;
 };
-// function hangmanImages() {
-//     let tries = Number(document.getElementById("tries").innerHTML);
-//
-//     let images = `assets/images/hangman/${tries}.png`;
-//     let id = document.getElementById("hanging");
-//
-//     id.setAttribute("src", images);
-//
-//     id.value;
-// }
