+++ assets/js/components/index.js
@@ -1,0 +1,2 @@
+export * from "./hangmanImages.js";
+export * from "./displayArrayToScreen.js";
