+++ assets/js/script.js
@@ -1,9 +1,8 @@
 import { names } from "./names.js";
 
 import { increase, decrease } from "./utils/scoring";
-import { hangmanImages } from "./utils/displayToScreen";
+import { hangmanImages, displayArrayToScreen } from "./components";
 
-// global variables
 let timer;
 
 // wait for the document to load then run starting functions
@@ -60,23 +59,6 @@ document.addEventListener("DOMContentLoaded", function () {
 });
 
 /**
- * passing in a param will print out the array to the screen were necessary
- * @param splitNameToArray
- */
-function displayArrayToScreen(splitNameToArray) {
-  let id = document.getElementById("letters");
-
-  for (let splitNameToArrayElement of splitNameToArray) {
-    let h3 = document.createElement("h3");
-    h3.setAttribute("id", "spacing");
-
-    h3.textContent = splitNameToArrayElement;
-
-    id.appendChild(h3);
-  }
-}
-
-/**
  * randomly selects a value from names array
  */
 function randomArrayValue() {
@@ -84,7 +66,7 @@ function randomArrayValue() {
 }
 
 /**
- * starts the timer on keydown event and sets the text content by id
+ * starts timer on keydown event in alphabet string
  * @param id
  */
 function startTimer(id) {
