--- assets/js/utils/displayToScreen/index.js
@@ -1,1 +1,0 @@
-export * from "./hangmanImages.js";
